+++ quantify/scheduler/compilation.py
@@ -1,17 +1,22 @@
 # -----------------------------------------------------------------------------
 # Description:    Compiler for the quantify.scheduler.
 # Repository:     https://gitlab.com/quantify-os/quantify-scheduler
-# Copyright (C) Qblox BV & Orange Quantum Systems Holding BV (2020)
+# Copyright (C) Qblox BV & Orange Quantum Systems Holding BV (2020-2021)
 # -----------------------------------------------------------------------------
+from __future__ import annotations
+from typing import TYPE_CHECKING
 import logging
 import jsonschema
-from typing import Callable
-from quantify.scheduler.types import Schedule
-from quantify.scheduler.pulse_library import ModSquarePulse, DRAGPulse, IdlePulse, SoftSquarePulse
+import importlib
+from quantify.scheduler.resources import ClockResource, BasebandClockResource
+from quantify.scheduler.pulse_library import SquarePulse, DRAGPulse, IdlePulse, SoftSquarePulse
 from quantify.utilities.general import load_json_schema
 
+if TYPE_CHECKING:
+    from quantify.scheduler.types import Schedule
 
-def _determine_absolute_timing(schedule, clock_unit='physical'):
+
+def determine_absolute_timing(schedule: Schedule, time_unit='physical'):
     """
     Determines the absolute timing of a schedule based on the timing constraints.
 
@@ -19,11 +24,11 @@ def _determine_absolute_timing(schedule, clock_unit='physical'):
     ----------
     schedule : :class:`~quantify.scheduler.Schedule`
         The schedule for which to determine timings.
-    clock_unit : str
+    time_unit : str
         Must be ('physical', 'ideal') : whether to use physical units to determine the
         absolute time or ideal time.
-        When clock_unit == "physical" the duration attribute is used.
-        When clock_unit == "ideal" the duration attribute is ignored and treated as if it is 1.
+        When time_unit == "physical" the duration attribute is used.
+        When time_unit == "ideal" the duration attribute is ignored and treated as if it is 1.
 
 
     Returns
@@ -40,7 +45,6 @@ def _determine_absolute_timing(schedule, clock_unit='physical'):
         3. determining the of the start of the operation based on the rel_time and duration of operations.
 
     """
-
     if len(schedule.timing_constraints) == 0:
         raise ValueError("schedule '{}' contains no operations".format(schedule.name))
 
@@ -61,24 +65,23 @@ def _determine_absolute_timing(schedule, clock_unit='physical'):
             ref_op = schedule.operations[ref_constr['operation_hash']]
 
         # duration = 1 is useful when e.g., drawing a circuit diagram.
-        duration_ref_op = ref_op.duration if clock_unit == 'physical' else 1
+        duration_ref_op = ref_op.duration if time_unit == 'physical' else 1
 
-        # determine
         if t_constr['ref_pt'] == 'start':
             t0 = ref_constr['abs_time']
         elif t_constr['ref_pt'] == 'center':
-            t0 = ref_constr['abs_time'] + duration_ref_op/2
+            t0 = ref_constr['abs_time'] + duration_ref_op / 2
         elif t_constr['ref_pt'] == 'end':
             t0 = ref_constr['abs_time'] + duration_ref_op
         else:
             raise NotImplementedError('Timing "{}" not supported by backend'.format(ref_constr['abs_time']))
 
-        duration_new_op = curr_op.duration if clock_unit == 'physical' else 1
+        duration_new_op = curr_op.duration if time_unit == 'physical' else 1
 
         if t_constr['ref_pt_new'] == 'start':
             t_constr['abs_time'] = t0 + t_constr['rel_time']
         elif t_constr['ref_pt_new'] == 'center':
-            t_constr['abs_time'] = t0 + t_constr['rel_time'] - duration_new_op/2
+            t_constr['abs_time'] = t0 + t_constr['rel_time'] - duration_new_op / 2
         elif t_constr['ref_pt_new'] == 'end':
             t_constr['abs_time'] = t0 + t_constr['rel_time'] - duration_new_op
 
@@ -98,15 +101,7 @@ def _find_edge(device_cfg, q0, q1, op_name):
     return edge_cfg
 
 
-def _walk_address(device_cfg, address):
-    paths = address.split(':')
-    curr_level = device_cfg['qubits']  # todo, make this work with more than just qubits
-    for path in paths:
-        curr_level = curr_level[path]
-    return curr_level
-
-
-def _add_pulse_information_transmon(schedule, device_cfg: dict):
+def add_pulse_information_transmon(schedule: Schedule, device_cfg: dict):
     """
     Adds pulse information specified in the device config to the schedule.
 
@@ -144,42 +139,55 @@ def _add_pulse_information_transmon(schedule, device_cfg: dict):
     validate_config(device_cfg, scheme_fn='transmon_cfg.json')
 
     for op in schedule.operations.values():
-        if 'operation_type' not in op['gate_info'] and len(op['pulse_info']) > 0:
-            # this is a pulse operation, make sure the address is resolved
-            for pulse in op.data['pulse_info']:
-                pulse['channel'] = _walk_address(device_cfg, pulse['channel'])
+        if op.valid_pulse:
+            for p in op['pulse_info']:
+                if p['clock'] not in schedule.resources:
+                    raise ValueError("Operation '{}' contains an unknown clock '{}'; ensure this resource has been "
+                                     "added to the schedule.".format(op.hash, p['clock']))
             continue
 
         if op['gate_info']['operation_type'] == 'measure':
             for q in op['gate_info']['qubits']:
-                q_cfg = _walk_address(device_cfg, q)
+                q_cfg = device_cfg["qubits"][q]
                 # readout pulse
-                if q_cfg['ro_pulse_type'] == 'square':
-                    op.add_pulse(ModSquarePulse(amp=q_cfg['ro_pulse_amp'],
-                                                duration=q_cfg['ro_pulse_duration'],
-                                                ch=q_cfg['ro_ch'],
-                                                freq_mod=q_cfg['ro_pulse_modulation_freq'],
-                                                t0=0))
+                if q_cfg['params']['ro_pulse_type'] == 'square':
+                    op.add_pulse(SquarePulse(amp=q_cfg['params']['ro_pulse_amp'],
+                                             duration=q_cfg['params']['ro_pulse_duration'],
+                                             port=q_cfg['resources']['port_ro'],
+                                             clock=q_cfg['resources']['clock_ro'],
+                                             t0=0))
                     # acquisition integration window
-                    op.add_pulse(ModSquarePulse(amp=1,
-                                                duration=q_cfg['ro_acq_integration_time'],
-                                                ch="{}_READOUT".format(q_cfg['ro_ch']),
-                                                freq_mod=-q_cfg['ro_pulse_modulation_freq'],
-                                                t0=q_cfg['ro_acq_delay']))
+                    acquisition = SquarePulse(amp=1,
+                                              duration=q_cfg['params']['ro_acq_integration_time'],
+                                              port=q_cfg['resources']['port_ro'],
+                                              clock=q_cfg['resources']['clock_ro'],
+                                              t0=q_cfg['params']['ro_acq_delay'])
+                    acquisition.mark_as_acquisition()
+                    op.add_pulse(acquisition)
+                    # add clock to resources
+                    if q_cfg['resources']['clock_ro'] not in schedule.resources.keys():
+                        schedule.add_resources(
+                            [ClockResource(q_cfg['resources']['clock_ro'], freq=q_cfg['params']['ro_freq'])])
 
         elif op['gate_info']['operation_type'] == 'Rxy':
             q = op['gate_info']['qubits'][0]
             # read info from config
             q_cfg = device_cfg['qubits'][q]
 
-            G_amp = q_cfg['mw_amp180']*op['gate_info']['theta'] / 180
-            D_amp = G_amp * q_cfg['mw_motzoi']
+            G_amp = q_cfg['params']['mw_amp180']*op['gate_info']['theta'] / 180
+            D_amp = G_amp * q_cfg['params']['mw_motzoi']
 
             pulse = DRAGPulse(
-                G_amp=G_amp, D_amp=D_amp, phase=op['gate_info']['phi'],
-                ch=q_cfg['mw_ch'],  duration=q_cfg['mw_duration'],
-                freq_mod=q_cfg['mw_modulation_freq'])
+                G_amp=G_amp, D_amp=D_amp,
+                phase=op['gate_info']['phi'],
+                port=q_cfg['resources']['port_mw'],
+                duration=q_cfg['params']['mw_duration'],
+                clock=q_cfg['resources']['clock_01'])
             op.add_pulse(pulse)
+
+            # add clock to resources
+            if q_cfg['resources']['clock_01'] not in schedule.resources.keys():
+                schedule.add_resources([ClockResource(q_cfg['resources']['clock_01'], freq=q_cfg['params']['mw_freq'])])
 
         elif op['gate_info']['operation_type'] == 'CNOT':
             # These methods don't raise exceptions as they will be implemented shortly
@@ -200,15 +208,21 @@ def _add_pulse_information_transmon(schedule, device_cfg: dict):
                 except ValueError:
                     raise
 
-            amp = edge_cfg['flux_amp_control']
-            pulse = SoftSquarePulse(amp=amp, duration=edge_cfg['flux_duration'], ch=edge_cfg['flux_ch_control'])
+            amp = edge_cfg['params']['flux_amp_control']
+
+            # FIXME: placeholder. currently puts a soft square pulse on the designated port of both qubits
+            pulse = SoftSquarePulse(amp=amp, duration=edge_cfg['params']['flux_duration'],
+                                    port=edge_cfg['resource_map'][q0], clock=BasebandClockResource.IDENTITY)
+            pulse = SoftSquarePulse(amp=amp, duration=edge_cfg['params']['flux_duration'],
+                                    port=edge_cfg['resource_map'][q1], clock=BasebandClockResource.IDENTITY)
+
             op.add_pulse(pulse)
         elif op['gate_info']['operation_type'] == 'reset':
             # Initialization through relaxation
             qubits = op['gate_info']['qubits']
             init_times = []
             for q in qubits:
-                init_times.append(device_cfg['qubits'][q]['init_duration'])
+                init_times.append(device_cfg['qubits'][q]['params']['init_duration'])
             op.add_pulse(IdlePulse(max(init_times)))
 
         else:
@@ -240,7 +254,8 @@ def validate_config(config: dict, scheme_fn: str):
     return True
 
 
-def qcompile(schedule: Schedule, device_cfg: dict, backend: Callable = None, **kwargs):
+def qcompile(schedule: Schedule, device_cfg: dict,
+             hardware_mapping: dict = None, **kwargs):
     """
     Compile and assemble a schedule into deployables.
 
@@ -249,9 +264,11 @@ def qcompile(schedule: Schedule, device_cfg: dict, backend: Callable = None, **k
     schedule : :class:`~quantify.scheduler.Schedule`
         To be compiled
     device_cfg : dict
-        Specifying the required pulse information. The device_cfg schema is specified in
-    backend : Callable
-        To the compiler, assembles the program(s).
+        Device specific configuration, defines the compilation step from
+        the gate-level to the pulse level description.
+    hardware_mapping: dict
+        hardware mapping, defines the compilation step from
+        the pulse-level to a hardware backend.
 
     Returns
     ----------
@@ -261,12 +278,29 @@ def qcompile(schedule: Schedule, device_cfg: dict, backend: Callable = None, **k
 
     .. rubric:: Configuration specification
 
-
     .. jsonschema:: schemas/transmon_cfg.json
+
+    .. todo::
+
+        Add a schema for the hardware mapping.
     """
-    schedule = _add_pulse_information_transmon(schedule=schedule, device_cfg=device_cfg)
-    schedule = _determine_absolute_timing(schedule=schedule, clock_unit='physical')
-    if backend:
-        return backend(schedule, **kwargs)
+
+    device_bck_name = device_cfg['backend']
+    if device_bck_name != 'quantify.scheduler.compilation.add_pulse_information_transmon':
+        raise NotImplementedError
+    (mod, cls) = (device_bck_name.rsplit(".", 1))
+    device_compile = getattr(importlib.import_module(mod), cls)
+
+    schedule = device_compile(schedule=schedule, device_cfg=device_cfg)
+    schedule = determine_absolute_timing(schedule=schedule, time_unit='physical')
+
+    if hardware_mapping is not None:
+        bck_name = hardware_mapping['backend']
+        # import the required backend callable to compile onto the backend
+        (mod, cls) = (bck_name.rsplit(".", 1))
+        # compile using the appropriate hardware backend
+        hardware_compile = getattr(importlib.import_module(mod), cls)
+        # FIXME: still contains a hardcoded argument in the kwargs
+        return hardware_compile(schedule, mapping=hardware_mapping, **kwargs)
     else:
         return schedule